add some updates. now need roster_resource status and then fg stats

+++ ulmg/management/commands/fix_dupes.py
@@ -9,7 +9,10 @@ from ulmg import models, utils
 
 
 class Command(BaseCommand):
-    def fix_fg_dupes(self):
+    def add_arguments(self, parser):
+        parser.add_argument("--trial", action="store_true", help="Explain the duplicates but don't delete anything")
+
+    def fix_fg_dupes(self, *args, **options):
         d1 = (
             models.Player.objects.exclude(fg_id__isnull=True).values('fg_id')
             .annotate(count=Count('id'))
@@ -18,27 +21,28 @@ class Command(BaseCommand):
             .filter(count__gt=1)
         )
 
-        print(f"initial duplicates: {d1.count()}")
-        for d in d1:
-            print(d)
-            objs = models.Player.objects.filter(fg_id=d['fg_id'])
-            # for o in objs:
-            #     # if not o.team:
-            #     #     o.delete()
+        print(f"initial FG duplicates: {d1.count()}")
 
-        d2 = (
-            models.Player.objects.exclude(fg_id__isnull=True).values('fg_id')
-            .annotate(count=Count('id'))
-            .values('fg_id')
-            .order_by()
-            .filter(count__gt=1)
-        ) 
+        if not options['trial']:
+            for d in d1:
+                objs = models.Player.objects.filter(fg_id=d['fg_id'])
+                for o in objs:
+                    if not o.team and not o.mlbam_id:
+                        o.delete()
 
-        print(f"remaining duplicates: {d2.count()}")
-        # for d in d2:
-        #     print(d)
+            d2 = (
+                models.Player.objects.exclude(fg_id__isnull=True).values('fg_id')
+                .annotate(count=Count('id'))
+                .values('fg_id')
+                .order_by()
+                .filter(count__gt=1)
+            ) 
 
-    def fix_mlbam_dupes(self):
+            print(f"remaining FG duplicates: {d2.count()}")
+            for d in d2:
+                print(d)
+
+    def fix_mlbam_dupes(self, *args, **options):
         d1 = (
             models.Player.objects.exclude(mlbam_id__isnull=True).values('mlbam_id')
             .annotate(count=Count('id'))
@@ -47,25 +51,27 @@ class Command(BaseCommand):
             .filter(count__gt=1)
         )
 
-        print(f"initial duplicates: {d1.count()}")
-        for d in d1:
-            objs = models.Player.objects.filter(mlbam_id=d['mlbam_id'])
-            for o in objs:
-                if not o.team and not o.fg_id:
-                    o.delete()
+        print(f"initial MLB duplicates: {d1.count()}")
 
-        d2 = (
-            models.Player.objects.exclude(mlbam_id__isnull=True).values('mlbam_id')
-            .annotate(count=Count('id'))
-            .values('mlbam_id')
-            .order_by()
-            .filter(count__gt=1)
-        ) 
+        if not options['trial']:
+            for d in d1:
+                objs = models.Player.objects.filter(mlbam_id=d['mlbam_id'])
+                for o in objs:
+                    if not o.team and not o.fg_id:
+                        o.delete()
 
-        print(f"remaining duplicates: {d2.count()}")
-        for d in d2:
-            print(d)
+            d2 = (
+                models.Player.objects.exclude(mlbam_id__isnull=True).values('mlbam_id')
+                .annotate(count=Count('id'))
+                .values('mlbam_id')
+                .order_by()
+                .filter(count__gt=1)
+            ) 
+
+            print(f"remaining MLB duplicates: {d2.count()}")
+            for d in d2:
+                print(d)
 
     def handle(self, *args, **options):
-        self.fix_mlbam_dupes()
-        # self.fix_fg_dupes()
+        self.fix_mlbam_dupes(*args, **options)
+        self.fix_fg_dupes(*args, **options)
